requirement.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
drf-yasg>=1.21.7
mysqlclient>=2.2.0
PyMySQL>=1.1.0
python-dotenv>=1.0.1
django-cors-headers>=4.3.1

